requirements.txt
langchain
openai
qdrant-client
tqdm
python-dotenv
requests